backend/requirements.txt
fastapi
uvicorn
minio
python-multipart
dotenv
google-auth
google-auth-oauthlib
google-api-python-client
pandas
openpyxl
scikit-learn
pymysql
requests
python-dotenv
motor
certifi
pydantic
pyarrow
starlette
pymongo
cryptography
pyspark
numpy
